pyproject.toml
[tool.poetry]
name = "kyouko"
version = "0.1.0"
description = "kyouko"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7.3"
nonebot2 = "^2.0.0-beta.1"

[tool.poetry.dev-dependencies]
nb-cli = "^0.6.0"

[tool.nonebot]
plugins = ["nonebot_plugin_status", "nonebot_plugin_tarot", "nonebot_plugin_covid19_news", "nonebot_plugin_memes", "nonebot_plugin_abstract", "nonebot_plugin_analysis_bilibili", "nonebot_plugin_autohelp", "nonebot_plugin_petpet"]
plugin_dirs = ["src/plugins"]

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
